python/requirements.txt
web3==6.11.3
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
cryptography==41.0.7
eth-account==0.9.0
eth-keys==0.4.0
hexbytes==0.3.1
